+++ speedy-murmurs-simulator/src/main/java/treeembedding/treerouting/TreerouteCPLPPP.java
@@ -16,9 +16,9 @@ public class TreerouteCPLPPP extends TreerouteNH {
 
   @Override
   protected double dist(int node, int neighbor, int dest) {
-    int[] a = this.coords[neighbor];
-    int[] b = this.coords[dest];
-    int[] own = this.coords[node];
+    long[] a = this.coords[neighbor];
+    long[] b = this.coords[dest];
+    long[] own = this.coords[node];
     int cpl = 0;
     int depth = sp.getDepth(neighbor);
     while (cpl < depth && cpl < b.length && a[cpl] == b[cpl]) {
